Add support for `LeaveType` in `LeaveDate` and `LeaveCalculate`, update salary calculation logic to use fractional leave days, and add new test cases for leave types.

+++ src/main/java/com/example/demo/Salary.java
@@ -2,14 +2,13 @@ package com.example.demo;
 
 public record Salary(int value, SalaryType salaryType) {
 
-    public int getSalaryActual(long shouldWorkDays, int leaveDays) {
+    public int getSalaryActual(long shouldWorkDays, double leaveDays) {
         if (shouldWorkDays == 0) {
             return 0;
         }
 
         double dailySalary = getDailySalary(shouldWorkDays);
-        long actualWorkDays = shouldWorkDays - leaveDays;
-
+        double actualWorkDays = shouldWorkDays - leaveDays;
 
         return (int) (dailySalary * actualWorkDays);
     }
